Feature: refactor print log resource
