+++ pkg/resource/rpc/log.go
@@ -1,0 +1,8 @@
+package rpc
+
+import (
+	"github.com/hashicorp/go-hclog"
+	"github.com/sev-2/raiden/pkg/logger"
+)
+
+var Logger hclog.Logger = logger.HcLog().Named("resource.rpc")
