+++ pkg/resource/tables/log.go
@@ -1,0 +1,8 @@
+package tables
+
+import (
+	"github.com/hashicorp/go-hclog"
+	"github.com/sev-2/raiden/pkg/logger"
+)
+
+var Logger hclog.Logger = logger.HcLog().Named("resource.tables")
