+++ pkg/state/state.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/sev-2/raiden"
@@ -71,12 +72,250 @@ type (
 
 		Through string
 	}
+
+	LocalState struct {
+		State      State
+		NeedUpdate bool
+		Mutex      sync.RWMutex
+	}
 )
 
 var (
 	StateFileDir  = "build"
 	StateFileName = "state"
 )
+
+func (s *LocalState) AddTable(table TableState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	s.State.Tables = append(s.State.Tables, table)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) FindTable(tableId int) (index int, tableState TableState, found bool) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	found = false
+
+	for i := range s.State.Tables {
+		t := s.State.Tables[i]
+
+		if t.Table.ID == tableId {
+			found = true
+			tableState = t
+			index = i
+			return
+		}
+	}
+	return
+}
+
+func (s *LocalState) UpdateTable(index int, state TableState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	s.State.Tables[index] = state
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) DeleteTable(tableId int) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	index := -1
+	for i := range s.State.Tables {
+		t := s.State.Tables[i]
+		if t.Table.ID == tableId {
+			index = i
+			break
+		}
+	}
+
+	if index == -1 {
+		return
+	}
+	s.State.Tables = append(s.State.Tables[:index], s.State.Tables[index+1:]...)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) AddRole(role RoleState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+	s.State.Roles = append(s.State.Roles, role)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) FindRole(roleId int) (index int, roleState RoleState, found bool) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	found = false
+
+	for i := range s.State.Roles {
+		r := s.State.Roles[i]
+
+		if r.Role.ID == roleId {
+			found = true
+			roleState = r
+			index = i
+			return
+		}
+	}
+	return
+}
+
+func (s *LocalState) UpdateRole(index int, state RoleState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	s.State.Roles[index] = state
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) DeleteRole(roleId int) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	index := -1
+	for i := range s.State.Roles {
+		r := s.State.Roles[i]
+
+		if r.Role.ID == roleId {
+			index = i
+			break
+		}
+	}
+
+	if index == -1 {
+		return
+	}
+	s.State.Roles = append(s.State.Roles[:index], s.State.Roles[index+1:]...)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) AddRpc(rpc RpcState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+	s.State.Rpc = append(s.State.Rpc, rpc)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) FindRpc(rpcId int) (index int, rpcState RpcState, found bool) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	found = false
+
+	for i := range s.State.Rpc {
+		r := s.State.Rpc[i]
+
+		if r.Function.ID == rpcId {
+			found = true
+			rpcState = r
+			index = i
+			return
+		}
+	}
+	return
+}
+
+func (s *LocalState) DeleteRpc(rpcId int) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	index := -1
+	for i := range s.State.Rpc {
+		r := s.State.Rpc[i]
+
+		if r.Function.ID == rpcId {
+			index = i
+			break
+		}
+	}
+
+	if index == -1 {
+		return
+	}
+	s.State.Rpc = append(s.State.Rpc[:index], s.State.Rpc[index+1:]...)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) UpdateRpc(index int, state RpcState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	s.State.Rpc[index] = state
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) AddStorage(storage StorageState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+	s.State.Storage = append(s.State.Storage, storage)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) FindStorage(storageId string) (index int, storageState StorageState, found bool) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	found = false
+
+	for i := range s.State.Storage {
+		r := s.State.Storage[i]
+
+		if r.Bucket.ID == storageId {
+			found = true
+			storageState = r
+			index = i
+			return
+		}
+	}
+	return
+}
+
+func (s *LocalState) UpdateStorage(index int, state StorageState) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	s.State.Storage[index] = state
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) DeleteStorage(storageId string) {
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+
+	index := -1
+	for i := range s.State.Storage {
+		r := s.State.Storage[i]
+
+		if r.Bucket.ID == storageId {
+			index = i
+			break
+		}
+	}
+
+	if index == -1 {
+		return
+	}
+	s.State.Roles = append(s.State.Roles[:index], s.State.Roles[index+1:]...)
+	s.NeedUpdate = true
+}
+
+func (s *LocalState) Persist() error {
+	s.Mutex.RLock()
+	defer s.Mutex.RUnlock()
+	if s.NeedUpdate {
+		if err := Save(&s.State); err != nil {
+			return err
+		}
+		s.NeedUpdate = false
+	}
+	return nil
+}
 
 func Save(state *State) error {
 	filePath, err := GetStateFilePath()
